requirements.txt
scipy
numpy
Pillow
opencv-python
pandas
